backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
# zoneinfo is stdlib in Python 3.9+; add backports.zoneinfo for 3.8
